+++ src/genom.h
@@ -366,7 +366,7 @@ extern POSTERS_INPUT_LIST * trouve_poster_input(char *name);
 extern DCL_NOM_STR * trouve_sdi_ref ( STR_REF_STR *ref );
 extern TYPE_STR * trouve_type ( TYPE_STR *t );
 extern DCL_NOM_STR * trouve_typedef ( TYPE_STR *t );
-extern int trouve_value ( char *symbole, int *pVal );
+extern int trouve_value ( char *symbole, long *pVal );
 extern void upCaseNames ( void );
 extern void upCaseArguments ( void );
 extern void * xalloc ( size_t n );
